pyproject.toml
[tool.poetry]
name = "tesla-cli"
version = "0.1.1"
description = ""
authors = ["MarcDufresne <marc.andre.dufresne@gmail.com>"]

packages = [
    {include = "tesla"}
]

[tool.poetry.dependencies]
python = "~3.6.8"
tesla-api = "^1.0"
requests = "^2.21"
click = "^7.0"
pyyaml = "^5.1"

[tool.poetry.dev-dependencies]
pyinstaller = "^3.4"
pytest = "^4.4"
pytest-sugar = "^0.9.2"

[tool.poetry.scripts]
tesla = 'tesla.cli:entry_point'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
